requirements.txt
streamlit
tensorflow-cpu
opencv-python-headless
gdown
pillow